Added functions to read in genomic data and negate it.
readFastaFile.py has two main functions so far. The first is to read in
data from a .fasta file. The genomic data we read in will be considered
the "positive strand" of DNA. The second function creates the "negative
strand" of DNA.

Also added in a test file called text.txt