Added two functions to help find potential genes
Added two functions
function 1 -- findPotentialGenes takes in the positive and negative
strand data and passes each strand individually to the second function.

function 2 - searchGene takes in one strand decided by function one,
and actually does all the process to find the potential genes. It
returns a named tuple which tells where the potential gene is on the
strand as well as waht contig it falls under.